[swf] Put all helper classes into renamed plugin class
Avoids conflicts with other plugins, e.g.

  superclass mismatch for class Identity
